Create CalcStrainCHM_sr01
Code creates a new mesh of the xy data - after removing wild vectors - of the PIV analysis and computes the corresponding xy position of the elements of the mesh. Further, the code removes the data within a predefined area which is not of interest (i.e. tunnel area). A new elements file is generated which neglects strain elements within the tunnel area.